chore: remove old upload endpoint
